+++ src/main/java/net/sourceforge/tess4j/Tesseract1.java
@@ -72,7 +72,7 @@ public class Tesseract1 extends TessAPI1 implements ITesseract {
         } finally {
             if (datapath == null) {
                 datapath = "./";
-            }            
+            }
         }
     }
 
@@ -517,7 +517,7 @@ public class Tesseract1 extends TessAPI1 implements ITesseract {
                     TessDeleteResultRenderer(renderer);
                 } catch (Exception e) {
                     // skip the problematic image file
-                    logger.error(e.getMessage(), e);
+                    logger.warn(e.getMessage(), e);
                 } finally {
                     if (workingTiffFile != null && workingTiffFile.exists()) {
                         workingTiffFile.delete();
@@ -535,6 +535,7 @@ public class Tesseract1 extends TessAPI1 implements ITesseract {
      * @param filename input file
      * @param renderer renderer
      * @throws TesseractException
+     * @return the average text confidence for Tesseract page result
      */
     private int createDocuments(String filename, TessResultRenderer renderer) throws TesseractException {
         TessBaseAPISetInputName(handle, filename); //for reading a UNLV zone file
@@ -583,7 +584,7 @@ public class Tesseract1 extends TessAPI1 implements ITesseract {
             return list;
         } catch (IOException ioe) {
             // skip the problematic image
-            logger.error(ioe.getMessage(), ioe);
+            logger.warn(ioe.getMessage(), ioe);
             throw new TesseractException(ioe);
         } finally {
             dispose();
@@ -629,32 +630,34 @@ public class Tesseract1 extends TessAPI1 implements ITesseract {
                 Word word = new Word(text, confidence, new Rectangle(left, top, right - left, bottom - top));
                 words.add(word);
             } while (TessPageIteratorNext(pi, pageIteratorLevel) == TRUE);
-
-            return words;
         } catch (Exception e) {
-            return words;
+            logger.warn(e.getMessage(), e);
         } finally {
             dispose();
         }
+        return words;
     }
 
     /**
-     * Creates documents and gives a result.
+     * Creates documents with OCR results.
      *
      * @param filenames array of input files
      * @param outputbases array of output filenames without extension
      * @param formats types of renderer
-     * @return Result with mean confidence and list of words with confidence for each word.
+     * @return OCR results
      * @throws TesseractException
      */
-    public Result createDocumentsWithResult(String[] filenames, String[] outputbases, List<ITesseract.RenderedFormat> formats) throws TesseractException {
+    @Override
+    public List<OCRResult> createDocumentsWithResults(String[] filenames, String[] outputbases, List<ITesseract.RenderedFormat> formats, int pageIteratorLevel) throws TesseractException {
         if (filenames.length != outputbases.length) {
             throw new RuntimeException("The two arrays must match in length.");
         }
 
         init();
         setTessVariables();
-        Result returnInfo = new Result(-1, null);
+
+        List<OCRResult> results = new ArrayList<OCRResult>();
+
         try {
             for (int i = 0; i < filenames.length; i++) {
                 File workingTiffFile = null;
@@ -668,13 +671,11 @@ public class Tesseract1 extends TessAPI1 implements ITesseract {
                     }
 
                     TessResultRenderer renderer = createRenderers(outputbases[i], formats);
-                    int conf =  createDocuments(filename, renderer);
-                    int pageIteratorLevel = TessPageIteratorLevel.RIL_WORD;
-                    returnInfo = new Result(conf, getCurrentWords(pageIteratorLevel));
+                    results.add(new OCRResult(createDocuments(filename, renderer), getRecognizedWords(pageIteratorLevel)));
                     TessDeleteResultRenderer(renderer);
                 } catch (Exception e) {
                     // skip the problematic image file
-                    logger.error(e.getMessage(), e);
+                    logger.warn(e.getMessage(), e);
                 } finally {
                     if (workingTiffFile != null && workingTiffFile.exists()) {
                         workingTiffFile.delete();
@@ -684,16 +685,17 @@ public class Tesseract1 extends TessAPI1 implements ITesseract {
         } finally {
             dispose();
         }
-        return returnInfo;
+
+        return results;
     }
 
     /**
-     * Gets recognized words at specified page iterator level for a page already loaded.
+     * Gets result words at specified page iterator level from recognized pages.
      *
      * @param pageIteratorLevel TessPageIteratorLevel enum
      * @return list of <code>Word</code>
      */
-    private List<Word> getCurrentWords(int pageIteratorLevel) {
+    private List<Word> getRecognizedWords(int pageIteratorLevel) {
         List<Word> words = new ArrayList<>();
 
         try {
@@ -718,11 +720,11 @@ public class Tesseract1 extends TessAPI1 implements ITesseract {
                 Word word = new Word(text, confidence, new Rectangle(left, top, right - left, bottom - top));
                 words.add(word);
             } while (TessPageIteratorNext(pi, pageIteratorLevel) == TRUE);
-
-            return words;
         } catch (Exception e) {
-            return words;
+            logger.warn(e.getMessage(), e);
         }
+
+        return words;
     }
 
     /**
