+++ src/main/java/net/sourceforge/tess4j/Word.java
@@ -18,7 +18,7 @@ package net.sourceforge.tess4j;
 import java.awt.Rectangle;
 
 /**
- * Encapsulates Tesseract OCR results.
+ * Encapsulates Tesseract OCR results at certain page iterator level.
  */
 public class Word {
